Split off p2p messages from shmsg
We defined all of the protoc generated code inside one package. This commit
moves the p2p related messages to a dedicated package.
