+++ .pre-commit-config.yaml
@@ -68,6 +68,7 @@ repos:
           (?x)(
               ^.*\.gen\.go$|
               ^rolling-shutter/shmsg/.*\.pb\.go$|
+              ^rolling-shutter/p2pmsg/.*\.pb\.go$|
               ^rolling-shutter/shcryptowasm/.*_wasm\.go$
           )
       - id: gofumpt
@@ -75,5 +76,6 @@ repos:
           (?x)(
               ^.*\.gen\.go$|
               ^rolling-shutter/shmsg/.*\.pb\.go$|
+              ^rolling-shutter/p2pmsg/.*\.pb\.go$|
               ^rolling-shutter/shcryptowasm/.*_wasm\.go$
           )
