+++ rolling-shutter/p2p/message.go
@@ -7,7 +7,7 @@ import (
 	"github.com/pkg/errors"
 	"google.golang.org/protobuf/proto"
 
-	"github.com/shutter-network/rolling-shutter/rolling-shutter/shmsg"
+	"github.com/shutter-network/rolling-shutter/rolling-shutter/p2pmsg"
 )
 
 // Message envelopes the serialized protobuf bytes with additional topic and sender info.
@@ -17,10 +17,10 @@ type Message struct {
 	SenderID string
 }
 
-func (msg Message) Unmarshal() (shmsg.P2PMessage, error) {
+func (msg Message) Unmarshal() (p2pmsg.P2PMessage, error) {
 	var err error
 
-	unmshl, err := shmsg.NewP2PMessageFromTopic(msg.Topic)
+	unmshl, err := p2pmsg.NewP2PMessageFromTopic(msg.Topic)
 	if err != nil {
 		return nil, errors.Wrap(err, "Failed to retrieve deserialisation type")
 	}
