+++ rolling-shutter/p2pmsg/doc.go
@@ -1,0 +1,3 @@
+package p2pmsg
+
+//go:generate protoc gossip.proto --go_out=./
