+++ rolling-shutter/p2pmsg/gossip.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
-package shmsg;
+package p2pmsg;
 
-option go_package = "./;shmsg";
+option go_package = "./;p2pmsg";
 
 message DecryptionTrigger {
     uint64 instanceID = 1;
